Version 8
Better GUI all around the app: implemented a new style for sliders, redesigned the main panel. The app should be able to render sharply on high resolution screen now, at least on Windows (if you enable override for high dpis, in javaw.exe's properties). Will make this a configuration in a future update.